Put run command cli reader into cli namespace
