+++ GPTifier/include/cli.hpp
@@ -13,6 +13,19 @@ struct ParamsEmbedding
     std::string model;
 };
 
+struct ParamsRun
+{
+    bool enable_export = true;
+    bool print_help = false;
+    std::string json_dump_file;
+    std::string model;
+    std::string prompt;
+    std::string prompt_file;
+    std::string temperature = "1";
+};
+
 bool get_opts_models(const int argc, char **argv);
 ParamsEmbedding get_opts_embed(const int argc, char **argv);
+ParamsRun get_opts_run(const int argc, char **argv);
+
 } // namespace cli
